feat: shared/ui 로직 분리
