+++ src/widgets/Header/index.ts
@@ -1,0 +1,1 @@
+export {Header} from './ui';
